Moved query functions from flow to datatools. Split out some repeated code into helper functions in utils
